new cmor_helpers file holds helper functions
cmor_mixer now creates xh and yh bnds. this was not supposed to take so long... why? see note-to-self...

...note to self, if you need to write data to a netcdf file... don't use [x][y][z] style indexing. you HAVE to use [x,y,z] for the assignment or the assignment fails.
