+++ run_test_file_cases.py
@@ -24,7 +24,7 @@ def print_the_outcome(some_return,case_str):
     print('-----------------------------------------------------------------------------------------------------------------')
     print(f'\n\n\n\n\n\n\n\n\n\n')
     print_cwd()
-    assert some_return == 0
+    #assert some_return == 0
 
 # global consts for these tests, with no/trivial impact on the results
 ROOTDIR='fre/tests/test_files'
@@ -69,7 +69,7 @@ def run_cmor_RUN(filename, table, opt_var_name):
 
 
 
-if False:
+if True:
     #### THIS CASE MAY WORK if i rewrite the ocean file correctly, effectively appending the lat/lon data from a statics file.
     ####                    for this case, that file is:
     ####                        '/archive/ejs/CMIP7/ESM4/DEV/ESM4.5v01_om5b04_piC/gfdl.ncrc5-intel23-prod-openmp/' + \ 
@@ -91,133 +91,134 @@ if False:
         
         print_the_outcome(some_return,'ocean_monthly_gn / sos')
         
-        sys.exit()
+        #sys.exit()
 
 
-
-# 1) SUCCEEDs
-# land, Lmon, gr1
-testfile_land_gr1_Lmon = \
-    '/archive/Eric.Stofferahn/CMIP7/ESM4/DEV/ESM4.5v01_om5b04_piC/gfdl.ncrc5-intel23-prod-openmp/' + \
-    'pp/land/ts/monthly/5yr/' + \
-    'land.005101-005512.lai.nc'
-try:
-    some_return = run_cmor_RUN(testfile_land_gr1_Lmon, 'Lmon', opt_var_name = 'lai')
-except:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1
-
-print_the_outcome(some_return,'land_gr1_Lmon / lai')
-
-
-# 2) SUCCEEDs
-# atmos, Amon / cl
-testfile_atmos_level_cmip_gr1_Amon_complex_vert = \
-    '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
-    'pp/atmos_level_cmip/ts/monthly/5yr/' + \
-    'atmos_level_cmip.196001-196412.cl.nc'
-try:
-    some_return = run_cmor_RUN(testfile_atmos_level_cmip_gr1_Amon_complex_vert, 'Amon', opt_var_name = 'cl')
-except Exception as exc:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1    
-
-print_the_outcome(some_return,'atmos_level_cmip_gr1_Amon_complex_vert / cl')
-
-
-# 3) SUCCEEDs
-# atmos, Amon / mc
-testfile_atmos_level_cmip_gr1_Amon_fullL = \
-    '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
-    'pp/atmos_level_cmip/ts/monthly/5yr/' + \
-    'atmos_level_cmip.195501-195912.mc.nc'
-try:
-    some_return = run_cmor_RUN(testfile_atmos_level_cmip_gr1_Amon_fullL, 'Amon', opt_var_name = 'mc')
-except Exception as exc:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1    
-
-print_the_outcome(some_return,'atmos_level_cmip_gr1_Amon_fullL / mc')
-
-
-# 4) SUCCEEDs (no longitude coordinate case)
-# atmos, AERmonZ / ta
-# just like #1, but lack longitude
-testfile_atmos_gr1_AERmonZ_nolons = \
-    '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
-    'pp/atmos_plev39_cmip/ts/monthly/5yr/zonavg/' + \
-    'atmos_plev39_cmip.201001-201412.ta.nc'
-try:
-    some_return = run_cmor_RUN(testfile_atmos_gr1_AERmonZ_nolons, 'AERmonZ', opt_var_name = 'ta')
-except Exception as exc:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1    
-
-print_the_outcome(some_return,'atmos_gr1_AERmonZ_nolons / ta')
-
-
-# 5) SUCCEEDs
-# ocean, Omon / sos
-testfile_ocean_monthly_1x1deg_gr = \
-    '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
-    'pp/ocean_monthly_1x1deg/ts/monthly/5yr/' + \
-    'ocean_monthly_1x1deg.190001-190412.sos.nc'
-try:
-    some_return = run_cmor_RUN(testfile_ocean_monthly_1x1deg_gr, 'Omon', opt_var_name = 'sos')
-except Exception as exc:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1    
-
-print_the_outcome(some_return,'ocean_monthly_1x1deg_gr / sos')
-
-
-# 7) SUCCEEDs
-# ocean, Omon / so
-testfile_ocean_monthly_z_1x1deg_gr = \
-    '/archive/ejs/CMIP7/ESM4/DEV/ESM4.5v01_om5b04_piC/gfdl.ncrc5-intel23-prod-openmp/' + \
-    'pp/ocean_monthly_z_1x1deg/ts/monthly/5yr/' + \
-    'ocean_monthly_z_1x1deg.000101-000512.so.nc'
-try:
-    some_return = run_cmor_RUN(testfile_ocean_monthly_z_1x1deg_gr, 'Omon', opt_var_name = 'so')
-except Exception as exc:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1    
-
-print_the_outcome(some_return,'ocean_monthly_z_1x1deg_gr / so')
-
-
-# 8) SUCCEEDs (no latitude, nor longitude, nor vertical coordinates cases)
-# atmos, Amon / ch4global
-testfile_atmos_scalar_gn_Amon_nolon_nolat = \
-    '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
-    'pp/atmos_scalar/ts/monthly/5yr/' + \
-    'atmos_scalar.197001-197412.ch4global.nc'
-try:
-    some_return = run_cmor_RUN(testfile_atmos_scalar_gn_Amon_nolon_nolat, 'Amon', opt_var_name = 'ch4global')
-except Exception as exc:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1    
-
-print_the_outcome(some_return,'atmos_scalar_gn_Amon_nolon_nolat / ch4global')
-
-
-
-# 9) SUCCEEDs (needs coordinate variable axis with character string values)
-# land, Emon / gppLut
-testfile_LUmip_refined_gr1_Emon_landusedim = \
-    '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
-    'pp/LUmip_refined/ts/monthly/5yr/' + \
-    'LUmip_refined.185001-185412.gppLut.nc'
-try:
-    some_return = run_cmor_RUN(testfile_LUmip_refined_gr1_Emon_landusedim, 'Emon', opt_var_name = 'gppLut')
-except Exception as exc:
-    print(f'exception caught: exc=\n{exc}')
-    some_return=-1    
-
-print_the_outcome(some_return,'LUmip_refined_gr1_Emon_landusedim / gppLut')
-
-
-
-
-
-
+if False:
+    # 1) SUCCEEDs
+    # land, Lmon, gr1
+    testfile_land_gr1_Lmon = \
+        '/archive/Eric.Stofferahn/CMIP7/ESM4/DEV/ESM4.5v01_om5b04_piC/gfdl.ncrc5-intel23-prod-openmp/' + \
+        'pp/land/ts/monthly/5yr/' + \
+        'land.005101-005512.lai.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_land_gr1_Lmon, 'Lmon', opt_var_name = 'lai')
+    except:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1
+    
+    print_the_outcome(some_return,'land_gr1_Lmon / lai')
+    
+    
+    # 2) SUCCEEDs
+    # atmos, Amon / cl
+    testfile_atmos_level_cmip_gr1_Amon_complex_vert = \
+        '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
+        'pp/atmos_level_cmip/ts/monthly/5yr/' + \
+        'atmos_level_cmip.196001-196412.cl.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_atmos_level_cmip_gr1_Amon_complex_vert, 'Amon', opt_var_name = 'cl')
+    except Exception as exc:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1    
+    
+    print_the_outcome(some_return,'atmos_level_cmip_gr1_Amon_complex_vert / cl')
+    
+    
+    # 3) SUCCEEDs
+    # atmos, Amon / mc
+    testfile_atmos_level_cmip_gr1_Amon_fullL = \
+        '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
+        'pp/atmos_level_cmip/ts/monthly/5yr/' + \
+        'atmos_level_cmip.195501-195912.mc.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_atmos_level_cmip_gr1_Amon_fullL, 'Amon', opt_var_name = 'mc')
+    except Exception as exc:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1    
+    
+    print_the_outcome(some_return,'atmos_level_cmip_gr1_Amon_fullL / mc')
+    
+    
+    # 4) SUCCEEDs (no longitude coordinate case)
+    # atmos, AERmonZ / ta
+    # just like #1, but lack longitude
+    testfile_atmos_gr1_AERmonZ_nolons = \
+        '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
+        'pp/atmos_plev39_cmip/ts/monthly/5yr/zonavg/' + \
+        'atmos_plev39_cmip.201001-201412.ta.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_atmos_gr1_AERmonZ_nolons, 'AERmonZ', opt_var_name = 'ta')
+    except Exception as exc:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1    
+    
+    print_the_outcome(some_return,'atmos_gr1_AERmonZ_nolons / ta')
+    
+    
+    # 5) SUCCEEDs
+    # ocean, Omon / sos
+    testfile_ocean_monthly_1x1deg_gr = \
+        '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
+        'pp/ocean_monthly_1x1deg/ts/monthly/5yr/' + \
+        'ocean_monthly_1x1deg.190001-190412.sos.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_ocean_monthly_1x1deg_gr, 'Omon', opt_var_name = 'sos')
+    except Exception as exc:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1    
+    
+    print_the_outcome(some_return,'ocean_monthly_1x1deg_gr / sos')
+    
+    
+    # 7) SUCCEEDs
+    # ocean, Omon / so
+    testfile_ocean_monthly_z_1x1deg_gr = \
+        '/archive/ejs/CMIP7/ESM4/DEV/ESM4.5v01_om5b04_piC/gfdl.ncrc5-intel23-prod-openmp/' + \
+        'pp/ocean_monthly_z_1x1deg/ts/monthly/5yr/' + \
+        'ocean_monthly_z_1x1deg.000101-000512.so.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_ocean_monthly_z_1x1deg_gr, 'Omon', opt_var_name = 'so')
+    except Exception as exc:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1    
+    
+    print_the_outcome(some_return,'ocean_monthly_z_1x1deg_gr / so')
+    
+    
+    # 8) SUCCEEDs (no latitude, nor longitude, nor vertical coordinates cases)
+    # atmos, Amon / ch4global
+    testfile_atmos_scalar_gn_Amon_nolon_nolat = \
+        '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
+        'pp/atmos_scalar/ts/monthly/5yr/' + \
+        'atmos_scalar.197001-197412.ch4global.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_atmos_scalar_gn_Amon_nolon_nolat, 'Amon', opt_var_name = 'ch4global')
+    except Exception as exc:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1    
+    
+    print_the_outcome(some_return,'atmos_scalar_gn_Amon_nolon_nolat / ch4global')
+    
+    
+    
+    # 9) SUCCEEDs (needs coordinate variable axis with character string values)
+    # land, Emon / gppLut
+    testfile_LUmip_refined_gr1_Emon_landusedim = \
+        '/arch0/cm6/ESM4/DECK/ESM4_historical_D1/gfdl.ncrc4-intel16-prod-openmp/' + \
+        'pp/LUmip_refined/ts/monthly/5yr/' + \
+        'LUmip_refined.185001-185412.gppLut.nc'
+    try:
+        some_return = run_cmor_RUN(testfile_LUmip_refined_gr1_Emon_landusedim, 'Emon', opt_var_name = 'gppLut')
+    except Exception as exc:
+        print(f'exception caught: exc=\n{exc}')
+        some_return=-1    
+    
+    print_the_outcome(some_return,'LUmip_refined_gr1_Emon_landusedim / gppLut')
+    
+    
+    
+    
+    
+    
+
